copied command encoder classes
