+++ remoting/dolphin-remoting-shared/src/main/java/com/canoo/dolphin/impl/commands/InterruptLongPollCommand.java
@@ -15,13 +15,14 @@
  */
 package com.canoo.dolphin.impl.commands;
 
-import org.opendolphin.RemotingConstants;
-import org.opendolphin.core.comm.SignalCommand;
+import org.opendolphin.core.comm.Command;
 
-public final class InterruptLongPollCommand extends SignalCommand {
+import static org.opendolphin.core.comm.CommandConstants.INTERRUPT_LONG_POLL_COMMAND_ID;
+
+public final class InterruptLongPollCommand extends Command {
 
     public InterruptLongPollCommand() {
-        super(RemotingConstants.INTERRUPT_LONG_POLL_COMMAND_NAME);
+        super(INTERRUPT_LONG_POLL_COMMAND_ID);
     }
 }
