+++ remoting/dolphin-remoting-shared/src/main/java/org/opendolphin/core/comm/AttributeMetadataChangedCommand.java
@@ -15,6 +15,10 @@
  */
 package org.opendolphin.core.comm;
 
+import static org.opendolphin.core.comm.CommandConstants.ATTRIBUTE_METADATA_CHANGED_COMMAND_ID;
+
+//Deprecated because the only type that can change is the qualifier. We should name it to QualifierChangedCommand or something
+@Deprecated
 public final class AttributeMetadataChangedCommand extends Command {
 
     private String attributeId;
@@ -24,9 +28,12 @@ public final class AttributeMetadataChangedCommand extends Command {
     private Object value;
 
     public AttributeMetadataChangedCommand() {
+        super(ATTRIBUTE_METADATA_CHANGED_COMMAND_ID);
+
     }
 
     public AttributeMetadataChangedCommand(final String attributeId, final String metadataName, final Object value) {
+        this();
         this.attributeId = attributeId;
         this.metadataName = metadataName;
         this.value = value;
