+++ remoting/dolphin-remoting-shared/src/main/java/org/opendolphin/core/comm/ChangeAttributeMetadataCommand.java
@@ -15,6 +15,10 @@
  */
 package org.opendolphin.core.comm;
 
+import static org.opendolphin.core.comm.CommandConstants.CHANGE_ATTRIBUTE_METADATA_COMMAND_ID;
+
+//Deprecated because the only type that can change is the qualifier. We should name it to QualifierChangedCommand or something
+@Deprecated
 public final class ChangeAttributeMetadataCommand extends Command {
 
     private String attributeId;
@@ -22,9 +26,11 @@ public final class ChangeAttributeMetadataCommand extends Command {
     private Object value;
 
     public ChangeAttributeMetadataCommand() {
+        super(CHANGE_ATTRIBUTE_METADATA_COMMAND_ID);
     }
 
     public ChangeAttributeMetadataCommand(final String attributeId, final String metadataName, final Object value) {
+        this();
         this.attributeId = attributeId;
         this.metadataName = metadataName;
         this.value = value;
