+++ remoting/dolphin-remoting-shared/src/main/java/org/opendolphin/core/comm/Command.java
@@ -31,33 +31,16 @@ public abstract class Command implements Serializable {
 
     private final String id;
 
-    public Command() {
-        this.id = null;
-    }
-
     public Command(final String id) {
         this.id = id;
     }
 
-    @Deprecated
     public String getId() {
-        if(id == null) {
-            return idFor(getClass());
-        }
-        return id;
-    }
-
-    //TODO: REMOVE THIS!!!!!!!!!! see https://github.com/canoo/dolphin-platform/issues/513
-    @Deprecated
-    private static String idFor(final Class commandClass) {
-        String id = commandClass.getSimpleName();
-        id = id.replace("Command", "");
-        id = id.replace("Notification", "");
         return id;
     }
 
     public String toString() {
-        return "Command: " + getId();
+        return "Command of type " + getId();
     }
 
 }
